requirement.txt
greenlet==3.2.4
playwright==1.55.0
pyee==13.0.0
typing_extensions==4.15.0
psycopg2-binary